// packages/engine/src/networking/interfaces/WorldState.d.ts
export interface NetworkInputInterface {
    /** ID of network. */
    networkId: number;
    /** Button input received over the network. */
    buttons: Array<{
        input: InputAlias;
        value: NumericalType;
        lifecycleState: LifecycleValue;
    }>;
    /** Axes 1D input received over the network. */
    axes1d: Array<{
        input: InputAlias;
        value: NumericalType;
        lifecycleState: LifecycleValue;
    }>;
    /** Axes 2D input received over the network. */
    axes2d: Array<{
        input: InputAlias;
        value: NumericalType;
        lifecycleState: LifecycleValue;
    }>;
    /** Viewport vector of the client. */
    viewVector: {
        x: number;
        y: number;
        z: number;
    };
}